P118-034 another test for stripping constants in gg
We already have one test for stripping constants in the generated globals
within a single compilation unit. This tests is for constants initialized
with calls to functions from other compilation units.

* testsuite/gnatprove/tests/P118-034__flow_gg_const/
Added; the "Ordinary_Constant" should not be in the generated globals,
because it is initialized using a function without variable inputs.

Change-Id: I04bbe61827220dd4e5c4a169610062b06db2c965