P118-034 decouple constants (with variable inputs) from globals
Detecting constants with variable inputs is part of the "global generation"
framework, but unlike the Global contract constants are *not* attached to
any entity. In particular, they appear "outside" of unit-level instances of
generic subprograms, so attaching them to any "scope" is semantically
ill-defined.

This change is meant to support improvements of gg for generic IN
parameters. Like the rest of the "gg" work, it feels too big to be
documented in the commit message.

Change-Id: I1d9c77dfa93f963a37fa64a43c43ff3e05dc835d