Q628-024 Do not issue warning for ghost procedure with no effect
The purpose of ghost procedures is in many cases to drive the automatic
proof of a property through ghost code, the so-called "auto-active"
verification. Thus, it is useless in general to issue a warning about
the fact that such procedures have no effect on outputs. We now avoid
issuing this warning in such a case.

* gnat2why/flow/flow.adb
(Flow_Analyse_CUnit): Do not issue a warning for ghost subprograms.

* testsuite/gnatprove/tests/M220-043__ghost/test.out
* testsuite/gnatprove/tests/red_black_trees/test.out
* testsuite/gnatprove/tests/string_search/test.out
Update outputs.

Change-Id: I60dd8bd4ecd7e746a37dc8df16a606e126a72462