Q629-007 Expand all visible parts of abstract states
An abstract state may contain parts which are visible from a unit
even if its refinement is not. These parts can come from the
private part of the unit if the body is not in SPARK, or from other
private child units of the same parent.
Instead of only expending consituents that are in the private part of
the unit with the abstract state, iterate over the Part_Of
constituents of the abstract state and expand those which have been
marked as in SPARK.

* gnat2why/flow/flow_utility.adb
(Expand_Abstract_State): If the refinement of an abstract state is
 not in SPARK, expand all its Part_Of constituents that are in SPARK.

Change-Id: Ibcbd3ca48d6f560edb54acc0ad7aad19d2123594