Q713-025 traverse expression of the Address aspect
Aspect Address, unlike other aspects, is not translated to a semantically
equivalent pragma, and like other aspects it is not attached to the tree. We
need to explicitly traverse its expression, which may contain references to
objects and calls.

In theory, we should do this for all aspects without semantically equivalent
pragmas that do not require static expression. In practice, Address seems to
be the only such an aspect.

* gnat2why/spark/spark_register.adb
(Process_Node): traverse expression of the Address aspect.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Node): mirror the above change (this is a known code duplication).

* testsuite/gnatprove/tests/Q713-025__flow_import/pixel.adb
* testsuite/gnatprove/tests/Q713-025__flow_import/test.opt
Extend test with an Address aspect on a subprogram and enable it.

Change-Id: I17f5c325bec5afa2e4e02129f01200c6fb1389bb