Q713-033 Automatically unroll static loops without loop invariant
Loop invariants are a known difficulty for beginners to use formal
program verification. At the same time, many loops written in exercises
and to familiarize oneself with the technology have static bounds and
a small range. This leads to an interesting possibility: unrolling loops
with static bounds and not too many iterations (current threshold is at
20 iterations maximum). This way, simple loops with static bounds are
proved without requiring a loop invariant.

We let the user decide if she wants to use this feature or not:
- by only unrolling FOR loops that have no loop (in)variant
- by disabling this feature when the switch --no-loop-unrolling is used

As visible from the testsuite, in the majority of cases, this feature
leads to better provability, on the loops without loop invariant.
In a few cases, the increase in size of formulas due to unrolling might
decrease provability. In such a case, the user can either insert a
loop invariant of True in her loop, or use the switch --no-loop-unrolling.

Besides beginners, this feature could also be convenient to explore
new algorithms, as the provers can be relied on to prove FOR loops with
low number of iterations (under 20), before the actual proof for a larger
value of iterations is attempted. Counterexamples also tend to be better
on the unrolled loops.

* common/gnat2why_args.adb, common/gnat2why_args.ads
(Max_Loop_Unrolling): New constant for max range of loop to unroll.
(No_Loop_Unrolling): New option to prevent loop unrolling.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Recognize loops with static bounds and no
 loop (in)variant, and unroll these loops instead of pretending that
 they have a default loop invariant of True.
(Unroll_Loop): New function to unroll a loop.

* gnatprove/configuration.adb, gnatprove/configuration.ads
New switch --no-loop-unrolling to prevent loop unrolling.

* gnatprove/gnatprove.adb
(Pass_Extra_Options_To_Gnat2why): Pass the new option.

* share/spark/help.txt
Update help message with new switch.

* docs/ug/en/source/assertion_pragmas.rst
* docs/ug/en/source/how_to_write_loop_invariants.rst
Document the new feature and switch.

* docs/ug/gnatprove_by_example/examples/loop_unrolling.adb
* docs/ug/gnatprove_by_example/examples/loop_unrolling.ads
* docs/ug/gnatprove_by_example/examples/simple_loops_unroll.adb
* docs/ug/gnatprove_by_example/results/counterex_unprovable.prove
* docs/ug/gnatprove_by_example/results/loop_unrolling.flow
* docs/ug/gnatprove_by_example/results/loop_unrolling.prove
* docs/ug/gnatprove_by_example/results/simple_loops_unroll.flow
* docs/ug/gnatprove_by_example/results/simple_loops_unroll.prove
Add support examples with outputs.

* testsuite/gnatprove/tests/PA20-065__gnatprove_help_output/test.out
* testsuite/gnatprove/tests/PA20-065__gnatprove_output/test.out
Update output depending on help message.

* testsuite/gnatprove/tests/K330-014__loop_entry/test.out
* testsuite/gnatprove/tests/KB04-004__quantified/data.ads
* testsuite/gnatprove/tests/KB04-004__quantified/test.out
* testsuite/gnatprove/tests/M301-012__array_of_iter/test.out
* testsuite/gnatprove/tests/M318-008__flow_generics/test.out
* testsuite/gnatprove/tests/M318-009__flow_quantified_expressions/test.out
* testsuite/gnatprove/tests/M318-028__flow_main_program/test.out
* testsuite/gnatprove/tests/M426-014__flow_tick_old_and_loop_entry/test.ou
* testsuite/gnatprove/tests/M613-039__expr_actions/test.out
* testsuite/gnatprove/tests/M809-005__float_basics/proof/sessions/*
* testsuite/gnatprove/tests/M823-017__fixed_points/test.out
* testsuite/gnatprove/tests/M906-011__flow_detect_*/test.out
* testsuite/gnatprove/tests/M917-032__spark_mode/test.out
* testsuite/gnatprove/tests/MB12-007__unbound_symbol/test.out
* testsuite/gnatprove/tests/N425-004__loop_frame/test.out
* testsuite/gnatprove/tests/N715-015__flow_oo_legal/test.out
* testsuite/gnatprove/tests/O513-003__justification/test.out
* testsuite/gnatprove/tests/O728-016__flow_loop_variables/test.out
* testsuite/gnatprove/tests/OA20-032__refinement/test.out
* testsuite/gnatprove/tests/OC26-001/test.out
* testsuite/gnatprove/tests/P630-025__loopinv/test.out
* testsuite/gnatprove/tests/P630-025__task_init/test.out
* testsuite/gnatprove/tests/Q328-034__mut_discr/test.out
* testsuite/gnatprove/tests/autopilot/test.out
Update output, as more checks are proved due to unrolling.

* testsuite/gnatprove/tests/M919-035__floating_point/dynamic_float.adb
* testsuite/gnatprove/tests/M919-035__floating_point/test.py
* testsuite/gnatprove/tests/O416-002__dyn_pred_in_loop/test.py
* testsuite/gnatprove/tests/P518-021__array_frame_condition/test.py
* testsuite/gnatprove/tests/P518-021__loop_frame_condition/test.py
* testsuite/gnatprove/tests/tetris/test.gpr
* testsuite/gnatprove/tests/tetris/test.py
Prevent loop unrolling on a few tests, either due to the purpose of the
test, or to preserve precise results (like on tetris).

* testsuite/gnatprove/tests/Q713-033__loop_unroll/binary_search.adb
* testsuite/gnatprove/tests/Q713-033__loop_unroll/binary_search.ads
* testsuite/gnatprove/tests/Q713-033__loop_unroll/langley.adb
* testsuite/gnatprove/tests/Q713-033__loop_unroll/lcp.adb
* testsuite/gnatprove/tests/Q713-033__loop_unroll/lcp.ads
* testsuite/gnatprove/tests/Q713-033__loop_unroll/loopstat.adb
* testsuite/gnatprove/tests/Q713-033__loop_unroll/test.out
* testsuite/gnatprove/tests/Q713-033__loop_unroll/test.py
* testsuite/gnatprove/tests/Q713-033__loop_unroll/types.ads
* testsuite/gnatprove/tests/Q713-033__tictactoe/game.adb
* testsuite/gnatprove/tests/Q713-033__tictactoe/test.out
* testsuite/gnatprove/tests/Q713-033__tictactoe/test.py
* testsuite/gnatprove/tests/Q713-033__tictactoe/tictactoe.adb
* testsuite/gnatprove/tests/Q713-033__tictactoe/tictactoe.ads
New tests.

Change-Id: I303ea4949c9b17411835ab34ddc3fab2b15ccb06