PA18-012 Up_Project globals while building cfg
This is a workaround to fix an issue raised by the costumer. The underlying
problem is related to the flow scope for generics. As the proper fix will
require probably a lot of work this is a solution we could use for now but
should be removed once a proper fix would be in place.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): up project globals to the current scope.

* testsuite/gnatprove/tests/OC08-044__refinement
Test output updated. Better error location.

* testsuite/gnatprove/tests/PA18-012__flow_scope_for_generic_subprogram
Test output updated. Spurious check coming from wrong globals disappears.

Change-Id: I767f55bf3a8803c13a4180780f5e7a28bfe60a7b