P805-023 Marking of deferred constants top-down
For top-down marking, processing that is currently done when encountering a
declaration needs to be moved to the processing of the entity, because in
top-down marking, the declaration may never be marked.

This commit does this for deferred constants.

* spark_definition.adb
(Mark): remove special treatment of deferred constants
(Mark_Entity): add special treatment of deferred constants

Change-Id: I28d658388743f0737febac64b7c9e72eab39e0b3