Q721-024 fix finding the protected type in component objects
When applying the Depends contract of protected subprograms we substitute
references to the protected type with references to a current objects.

To do this we need to know the protected type, but this was only working for
standalone protected objects and not for protected components in arrays or
records. Now fixed.

* gnat2why/flow/flow_utility.adb
(Get_Depends): the protected type is always the scope of the protected
 subprogram; previously we were getting the type of the object, but this
 doesn't work for protected components in arrays and records. Remove
 assertion, that now is obvious (it is equivalent to condition of an IF
 statement that encloses this code).

* testsuite/gnatprove/tests/Q721-024__flow_po_depends/
Enable; add output.

Change-Id: I393b55fda8841f585f10f44b954aff19b46be5f0