Q724-019 Overridden primitive equality on record derivation
For record types and subtypes, a placeholder is defined for the
primitive equality so that it can be called easily from predefined
equality of other composite types. At the type completion, an axiom
is supplied whenever a user-provided primitive equality has been
found.
We used to generate the symbol for primitive equality in the
representative module of record types, which could be shared in a
record type hierarchy when no component was added by a derivation
(untagged or tagged with null record). As a result, we could have
several conflicting defining axioms for the same symbol if
equality was overridden in such a case.
Move the symbol for primitive equality to the specific record
modules.

* gnat2why/why/why-gen-records.adb
(Declare_Ada_Record): Declare abstract symbol for primitive
 equality.
(Declare_Rep_Record_Type): Do not declare abstract symbol for
 primitive equality.
(Oldest_Parent_With_Same_Fields): Fix typo, we should call
 Has_Private_Part to know whether a type has private parts and not
 Has_Private_Type.

Change-Id: I8cebe8de85f67cc938529ba5a6a447cc56c2033a