Q728-010 Check locations added as extra info for GPS plugin
This commit adds information to the JSON generated informations passed to
spark GPS plugin so that a messages now 'contains' the location of the
check it is associated to. Before, the location provided was the one of
the first failing assumption of the check, it is still the case but an
invisible location is added which is the location of the check (it is
necessary for manual proofs).
This change should not change any test. It is associated to a commit on
the spark plugin for GPS.

* gnat2why/flow/flow_error_messages.adb
(Add_Json_Msg): Adding the source location of the check VC_Loc. Adding
fields "check_file", "check_line" and "check_col" which are used to locate
the VC.
(Error_Msg_Proof): Adding the node associated to the VC (which contains the
location): VC_Loc.
Adding VC_Slc, the location generated from VC_loc.

* gnat2why/why/gnat2why-error-messages.adb
(Emit_proof_result): Adding appropriate VC_loc.



Change-Id: If3e3136bd39e6dd315d8d88bf2b81445761359c1