PC06-008 Add test
Add a test for this TN so we will notice when the issue will be solved (I
have already checked that the spurious error will be gone with Piotr's
patch for Q127-012).

* testsuite/gnatprove/tests/PC06-008__flow_generic_actual
Add test. The output currently issue a wrong check. This will be gone when
we will fix the handling of generic actuals.

Change-Id: I977daea2b6943de043ecb2b060b05f159f5acb9d