P819-010, PB07-037 fix handling of package locals
This is a substantial rewrite of how the package "local" variables are
handled, i.e. how they appear in the CFG and how they go into the generated
Initializes contract.

In particular, abstract states of the currently analyzed package no longer
appear in the CFG, because this caused aliasing with their constituents.
Thus, only constituents and "implicit" abstract states (i.e. variables and
constants with variable inputs declared in the private part of a package
declaration and in the package body declaration) appear on the CFG and are
registered already when building the graph, which avoids rediscovering them
with a yet another AST traversal in slicing.

* gnat2why/flow/flow.ads
(Flow_Global_Generation_Info): add containers for package-local variables.

* gnat2why/flow/flow-slice.ads
* gnat2why/flow/flow-slice.adb
(Compute_Globals): remove discovery of packge-local variables from slicing.

* gnat2why/flow/flow-control_flow_graph.adb
(Register_Own_Variable): new routine for detection of package-level
 variables; it is called in Do_Package_Declaration (to register constants
 and variables) and in Do_Package_Declaration (to register abstract states
 of nested packages).

(Do_Package_Declaration): respect SPARK_Mode => Off on a private part of
 the package specification; only traverse that part when no abstract states
 are declared for a package.

(Create): do not create vertices for abstract states of the current package,
 because they alias with their own constituents; respect SPARK_Mode => Off
 on a private part of the package specification.

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Is_Package_State): constants with variable inputs are also package state
 variables (except generic actual parameters); this fix is needed for
 Register_Own_Variable.

* gnat2why/flow/flow_generated_globals-partial.adb
* gnat2why/flow/flow_generated_globals-phase_2.adb
Various adaptations, just to keep the generation of Initializes working;
this needs to be revisited later anyway, so is not documented here.

* testsuite/gnatprove/tests/O909-044__flow_wrong_gen_initializes
This is a regression: hierarchical packages with hierarchical abstract
states currently requires that the Initializes contract of the nested
package is available when building the CFG of the enclosing package.

* testsuite/gnatprove/tests/O909-059__part_of
Remove spurious warning about the lack of procedure that can initialize
state whose refinement is null.

* testsuite/gnatprove/tests/P819-010__flow_locals
Remove spurious warning about no effect when initializing a variable that
is not even visible (because it is declared in a private part of a package
with explicit abstract states). Update comments in test. Now the CFGs for
this code look fine.

* testsuite/gnatprove/tests/PB07-037__flow_crash_nested_nested_package
Enable test; no longer crashes.

* testsuite/gnatprove/tests/Q424-015__import_state
New legitimate warning: there is no procedure that could initialize an
abstract state. Previously this warning was not emitted, because the state
has null refinement, but this fact should not be known to GNATprove, as the
refinement is hidden behind a SPARK_Mode => Off.

* testsuite/gnatprove/tests/TU__initial_condition_illegal
Still rejected (as expected), but with a better error message.

Change-Id: Icd2cc20901d0d509f995899d8f6ac7678fe8d2ff