P926-007 Remove outdated comment and non necessary Global => null
The work under P531-027 created some regressions that are to be fixed under
this ticket. In particualar a bunch of Global => null were added to some
tests but are no longer needed. Piotr already removed an outdated comment
under P531-027 (for test P819-010) and here I am removing the same outdated
comment for other two tests. There are still few remaining regressions to
be fixed under this ticket.

* testsuite/gnatprove/tests/N108-029__flow_generate_initializes_1/p.adb
* testsuite/gnatprove/tests/NC18-019__flow_nested_package_bodies/outer.ads
Removed the null global annotation as it is no longer necessary

Change-Id: I55d805e4ff44ab25437f985dc7a5fbb0cbe679bc