OB06-004 replicate test for phase-1-only gg
In the original test the problem happens in the GG phase 2, which relies
mostly on Entity_Names are thus is difficult to debug and to fix. This
commit replicates the problem within a single compilation unit, so it can
be debugged in phase 1, which relies exclusively on Entity_Ids.

* testsuite/gnatprove/tests/OB06-004__mutually_recursive_protected_procs/
Add code that replicates the problem within a single compilation unit;
update output.

Change-Id: I532a3891b3c37456c02f707ea588887199a5e98f