Q124-001 minimize outputs for caching
Update z3 and cvc4 submodules with commits that disable prover statistics
excepts "steps", which is needed for the --reply mode.

Similar update for gnatwhy3 to disable timing output except when called with
the --debug switch. Also includes a commit that should mangle the "rem" name
when generating an SMT-LIB file for Z3 (a small issue reported on this TN).

Change-Id: I904fbcad01f0212dad608fb3c1e0bdb729a91dce