PC09-011 Fix crash on null abstract state with no refinement
We were crashing on a package with a null abstract state and no refinement
as we were always checking for a Refined_State. In case of a null abstract
state we now do not look for state variables not mentioned in the
Refined_State. The check of not having (hidden) state variables in
presence of a null abstract state is carried out by the front end.

* gnat2why/flow/flow-analysis.adb
(Find_Hidden_Unexposed_State): do not check for unreferenced state
variables in case of a package with a null abstract state.

* testsuite/gnatprove/tests/PC09-011__flow_null_abstract_state
Enable test as no longer crashes.

Change-Id: I2971ff3bd86ca4b9e0dcc0e5d40ab76026c3ccab