Q824-003 Remove workaround for PA18-012
Under PA18-012 I did introduce a workaround to paper over an issue with
visibility of generic instances. It is no longer needed as this has been
fixed under Q329-005. With a new test just added by Piotr (Q821-003) and
the test under this ticket (Q824-003) we also saw that this workaround was
not correct.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): remove workaround that up projects globals while building cfg.

* testsuite/gnatprove/tests/OC08-044__refinement
Update test output. Error locations are not that good but this is unrelated
to the problem this workaround was willing to attack.

* testsuite/gnatprove/tests/Q821-003__flow_child
Enable test as no longer crashes. The output is currently not correct as
there is a spurious high check on a missing global.

* testsuite/gnatprove/tests/Q824-003__flow_child_of_private_child
Output updated. Wrong high-check no longer emitted.

Change-Id: I030a19fcd3a365a95296512d0628a6a4af56edb1