Q719-004 improve bugbox location info for flow and marking
Set Current_Error_Node at major source analysis routine, so that if we crash
then the bugbox at least an approximates the problematic source location.

* gnat2why/spark/spark_definition.adb
(Mark): set Current_Error_Node to the currently analyzed node.

* gnat2why/flow/flow-control_flow_graph.adb
(Process_Statement): same as above (for fine-grained info).

* gnat2why/flow/flow.adb
(Flow_Analyse_Entity): same as above (for coarse-grained info).

* gnat2why/flow/flow_generated_globals-partial.adb
(Preanalyze): same as above, for the intra-procedural part of the GG.
(Do_Global):  same as above, for the inter-procedural part of the GG.

Change-Id: If7e464cbf5ee98132b33e2b0de7790d2c913aa59