Q613-018 Fix implementation of Is_Boundary_Subprogram_For_Type
We now decide if a subprogram is a boundary subprogram for a type based on
its definition given in SPARK RM 7.3.2(1).

* gnat2why/flow/flow_refinement.adb
(Is_Boundary_Subprogram_For_Type): new implementation that checks if the
subprogram is declared within the immediate scope of the type and if it is
visible from outside the immediate scope of the type.

* testsuite/gnatprove/tests/Q613-018__flow_type_invariants
Add test

* testsuite/gnatprove/tests/P801-003__flow_type_invariants
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects
* testsuite/gnatprove/tests/red_black_trees
Update test outputs. Spurious high checks about calls from a type invariant
to a boundary subprogram are now no longer emitted.

Change-Id: I37773ebf23bb50e2a80006a908000232c83ccd16