Q329-005 simplify getting Flow_Scope of expression functions
The special-case for expression functions seems to better fit into existing
handling of Prev_Context, that is used to decide where the up-traversal came
from (i.e. either from private or visible package declarations). Now we
similarly use it to decide if we came from a body of an expression function.

Also, this patch brings us closer to only returning from Get_Flow_Scope from
contexts that are indeed Flow_Scopes (and functions are not) and applying
the generic-related machinery (in Generic_Parent_Or_Parent) only to generic
packages. Without it, it would be too difficult to handle generic expression
functions declared in package specs, e.g.:

  package P is
     generic function F return Boolean;
     function F return Boolean is (True);
  end P;

Semantically, the code behaves exactly as before (at least as far as our
testsuite is concerned).

* gnat2why/flow/flow_refinement.adb
(Get_Flow_Scope): move handling of expression functions from the code for
 N_Subprogram_Body to the code for N_Package_Specification.

Change-Id: Iae2c586ee5cf6d9726dab92c0ed1c12712af2648