Q908-011 Detect missing constituent
When checking consistency of the refinement of a package with Part_Of
constituents declared in private descendants we where missing
inconsistencies in case the Part_Of was declared not in the immediate child
but in a child of a child. This is now fixed.

* gnat2why/flow/flow-analysis.adb
(Check_Consistent_AS_For_Private_Child): check refinement of the package
where the encapsulating state is declared.

* testsuite/gnatprove/tests/Q908-011__flow_child_of_private_child
Update test output with correct detection of inconsistent refinement

Change-Id: I79a7ed61758ec76b358ddfc811d232b3347a33db