remove useless calls to Is_Object
In "Is_Object (E) and then Ekind (E) = E_Constant" the first conjunct is not
needed, because it is already implied by the second one (and it doesn't
protect against any run-time error, because Is_Object is just a wrapper for
Ekind).

* gnat2why/why/gnat2why-expr.adb
(Assume_Declaration_Of_Entity): remove useless code.

(no-tn-check)

Change-Id: Icc13cf92d9d9e051ba4358454c6b26bd3f975b7a