Q522-005 Update manual proof example in UG
CVC4 can now prove the example on which manual proof is based. Force the
use of Alt-Ergo alone on this test to make the example relevant.

* docs/ug/gnatprove_by_example/Makefile.examples
* docs/ug/gnatprove_by_example/examples/altergo.gpr
* docs/ug/gnatprove_by_example/examples/altergo.txt
Update with a special project for units to be analyzed with --prover=altergo
(only one test for now).

* docs/ug/gnatprove_by_example/manual_proof.rst
Update description of manual proof example.

* docs/ug/gnatprove_by_example/results/nonlinear.prove
Update output.

Change-Id: Ia18d282d39f4598d0a8e1a67b6af20094ecefce2