remove suspicious use of Get_Enclosing_Body_Flow_Scope
Routine Get_Enclosing_Body_Flow_Scope probably should never be exposed from
the Flow_Refinement package. Its only use in Do_Package_Body_Or_Stub was
undocumented and suspicious.

I see no reason why an Initializes aspect of a nested package should be
parsed in the scope where the *body* of that package is defined (and that
body might be not even visible from the outer package; certainly it is not
visible when the outer package body is hidden behind SPARK_Mode => Off).

The results of Parse_Initializes with the scope of the Initializes aspect
itself are the same as before, at least across our testsuite.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Package_Body_Or_Stub): parse Initializes aspect of nested packages with
 visibility of that aspect itself.

* gnat2why/flow/flow_refinement.ads
(Get_Enclosing_Body_Flow_Scope): move to package body; this routine is now
 only internally used by Is_Visible.

(no-tn-check)

Change-Id: If1a8543df6d502e30f4f25d55b8e22accf0727fe