Q329-005 use generic scope for instances of generic packages
This patch removes accesses to the components of the Flow_Scope from outside
of the Flow_Refinement package.

It also changes the scope where the Initializes aspect of instances of
generic packages is resolved: originally it was the scope of the instance,
but after recent work on visibility and generics, it should be the scope of
the generic itself. Now it is.

No tests affected (either we have too few tests with tricky Initializes on
instances of generic packages or this doesn't matter due to other problems
with generated Initializes).

* gnat2why/flow/flow_generated_globals-phase_2.adb
(GG_Get_Initializes): replace hand-made Flow_Scope with Get_Flow_Scope.

Change-Id: I48f4dfd476ee15193c70efbcb5bf2d53e22f5072