Q914-023 fix Get_Enclosing_Body_Flow_Scope for expression functions
For expression functions declared in package specs we pretend that their
bodies are defined in the package bodies. However, there might be no such
package bodies (e.g. because the body is not needed, or does not yet exists
or has been excluded from the analysis in the .gpr file).

This patch rewrites Get_Enclosing_Body_Flow_Scope to not rely on the
up-traversal of the AST of the potentially non-existing package bodies.

No tests affected (and writing a test where this matters would be hard).

* gnat2why/flow/flow_refinement.adb
(Get_Enclosing_Body_Flow_Scope): rewritten.

Change-Id: I9f9172ce3771e6c29cf34e1f1627e100e5979302