Q927-028 remove always true tests in Up_Project
Testing if the encapsulating state of a constituent is indeed an abstract
state doesn't make sense, as this is always the case. Most likely this test
was only needed at some early stage of the new GG branch.

* gnat2why/flow/flow_refinement.adb
(Up_Project): remove checks for being an abstract state; all versions (for
 Entity_Id, Entity_Name and Flow_Id).

Change-Id: If4007abd0eeb1b82dd589f7e9825a18c67d1b241