Q927-028 remove extra variant of Up_Project
We already have Up_Project routines for sets of Entity_Ids, Entity_Names and
Flow_Ids, which ideally should be refactored (possibly using a generic
instantiated with packages, which so far has not been used in gnat2why).

To keep the code as uniform as possible, we don't want a yet another variant
specialized for single Flow_Ids; instead, we can reuse the existing one by
calling it with a singleton set and knowing that results are also singleton.

* gnat2why/flow/flow_refinement.adb
(Up_Project): variant for single Flow_Ids removed.

(Up_Project): variant for dependency maps rewritten to use singleton sets.

Change-Id: I8274dc22bf0d8c48abef92cb6fd3f83c2d007034