Q918-021 disable test that randomly hangs the testsuite
This test should crash gnat2why due to infinite recursion, but randomly the
gnat2why process hangs and needs to be killed either manually (which is
annoying) or by the test driver (which happens after very long timeout).

We could either set some small timeout for this test, but that would be
guessing and could disrupt routine monitoring of the testsuite performance
by developers who run it with 'time' utility, or simply disable the test for
now. This patch does the latter as it is unlikely to solve the underlying
problem without noticing.

* testsuite/gnatprove/tests/Q510-009__flow_circular_global/test.opt
Disable test.

Change-Id: I590ceb0e8b7d903726b507a3bce6d3579cdb7ed9