Q927-028 revert "slight refactoring" of Is_Visible
The refactoring of Is_Visible, which occurred in the commit below, made the
code very hard to understand (as demonstrated by Q329-005, which introduced
Looking_From_Generic that had to be reverted).

This brings Is_Visible to state from before the refactoring, just with an
indeed slight modification, i.e. removal over-used local Context variable
from the "Check if Target_Scope is nested within S" section.

After this change the results of calls to Is_Visible agree with the old
code, at least within out testsuite.

commit 31d20619896419e0ba7fa0cc7ed978fd8226b634
Author: Florian Schanda <florian.schanda@altran.com>
Date:   Fri Aug 22 14:18:52 2014

    N711-045 flow - support for Default_Initial_Condition

    A few small changes after reviewing this work

    ...
    * Slightly refactor Is_Visible and add more comments
    ...

Change-Id: I9ef0fa45541c33defa9e8a9d0e863be9eebbe90c