minor refactoring to simplify debugging
Code that first collects something into a container and then uses that
container is difficult to debug. If something goes wrong we can't just look
at the backtrace, but need to break when the problematic value was added
into the container.

With this patch we process globals as we discover them in the Initializes
contract, without keeping them in a temporary container.

* gnat2why/flow/flow-control_flow_graph.adb
(Create): simplify picking globals from the Initializes contract.

(no-tn-check)

Change-Id: Ic9624ffbed4b9e2a3238ac56b83be453f0d09e8b