P609-046 remove duplicated condition for inlining globals
Deciding whether to "inline" global effects was explicit for procedure and
entry calls (in Do_Call_Statement, which is part of the CFG construction),
but implicit for function calls (in Do_Subprogram_Call, which is part of
Get_Variables), where it was delegated to Get_Globals.

This was confusing, impossible to maintain and inefficient. Now the explicit
decision is removed and only described in a comment.

No impact on the testsuite (intentionally).

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Call_Statement): remove unnecessary code that duplicated logic already
 implemented in Get_Globals.

Change-Id: Ie6867cf84f95dcb7582983c06e6f93bfad6988f5