Q920-019 adapt test after frontend fix
Frontend now correctly rejects references to objects and functions declared
in the private part of package spec from its Initial_Condition and
Initializes contracts.

* testsuite/gnatprove/tests/TU__initializes_illegal/test.out
Update error message; this test was meant to be illegal from the start.

* testsuite/gnatprove/tests/MA29-016__flow_pre_post/
Adapt test, as otherwise it would be now rejected by the frontend; adjust
line numbers. (This test was added with a commit message:

  "Analyze the initial condition in the correct scope (and make sure the
   variables seen in the spec are projected appropriately)."

so apparently it was an unintentional mistake to reference the private
variable in Initial_Condition).

Change-Id: Ie9953efd84d45842c43b27f426e483cbd5a1b86a