Q927-028, Q329-005 Revert change in Is_Visible
While working on visibility and generics I have introduced a wrong fix in
Is_Visible that returned False every time we ask if something is visible
from a generic unit. I am now reverting this change but this causes two
regressions in N814-006 and N814-007 that are related to a still existing
bug in Is_Visible. This will be fixed under Q927-028.

* gnat2why/flow/flow_refinement.adb
(Is_Visible): delete code that return False every time we are checking
visibility from a generic unit.

* testsuite/gnatprove/tests/N814-006__flow_spurious_ineffective_init
* testsuite/gnatprove/tests/N814-007__refined_post
These are two regressions. The high checks shouldn't be there. This is due
to a bug in Is_Visible that thinks that List_Handler.List is visible from
the generic itself.

* testsuite/gnatprove/tests/Q329-005__flow_visibility_generics
Enable test as no longer crashes.

* testsuite/gnatprove/tests/P715-019__flow_generic_termination/
Revert to previous output (which seems fine).

Change-Id: I20f67f04d6bc3f878429afa9ee7c837b02a2617c