Q927-025 Update itp__sumrange
Tactic induction does not take the same argument as before (an
optional argument is now possible). So we need to update the session
files accordingly (session files contains arguments of
transformation). This change only impacted itp__sumrange because it is
the only itp proof currently.

* testsuite/gnatprove/tests/itp__sumrange/proof/sessions/* :
Update session and remove shapes and backup files

* testsuite/gnatprove/tests/itp__sumrange/test*
Update tests and remove XFAIL

Change-Id: Ie35f978e982a9c16c70cc6246ec091cb99a58791