Q927-025 Change itp_server_test
This commit changes the way this test is run. We now send a request to
the gnat_server so that it exits. So, we don't stop it abruptly
after x seconds. This is more robust. Also, this removes the
timings.

* testsuite/gnatprove/itp__server_test/test.in
Reorder and add an exit request

* testsuite/gnatprove/itp__server_test/test.opt
Removed

* testsuite/gnatprove/itp__server_test/test.out
Updated

* testsuite/gnatprove/itp__server_test/test.py
Removed timing information.

Change-Id: Iab09a5a7279738ced2013b340852fb8bf7c8b4f3