Q825-009 Fix crash in internal protected call with Depends
Flow was crashing when analysing an internal call to a protected subprogram
with Depends. This is because we only created implicit parameter vertexes
for external calls (i.e. when we have a protected object) but not
otherwise. We are now creating implicit parameter vertexes also for
protected types even when the call is internal and we don't have an object
of the protected type.

* gnat2why/flow/flow-control_flow_graph.adb
(Process_Call_Actuals): add vertexes also when the protected call is
internal.

* gnat2why/flow/flow-interprocedural.adb
(Add_Simple_Procedure_Dependency): include also the protected type implicit
parameter as a global when we don't have a protected object.

* gnat2why/flow/flow_utility.adb
(Get_Depends): Substitute instances of the protected type with instances
of the protected object only when the call is external and therefore we
have an object (in particular, this does not happen when the call is
internal and therefore we will still have references to the protected
type).

* testsuite/gnatprove/tests/Q825-009__flow_protected
Enable test as no longer crashes. Add test output.

Change-Id: I3bba0bf67508937491ec77e1504f4a38a79591af