QA04-034 Fix crash on handling of calls to volatile functions
Calls to volatile functions were incorrectly translated in Why3, which
could cause a crash. The part of the context which defines temporary
variables, when a parameter is of a volatile type that needs
reconstruction after the call, was missing, which caused these variables
to be unknown in Why3. This is now fixed.

* gnat2why/why/gnat2why-expr.adb
(Insert_Ref_Context): Take into account a call to a volatile function.
(Transform_Function_Call): Allow references to appear during the
 translation of a call to a volatile function.

* gnat2why/why/why-atree-sprint.adb
(Print_Binding_Ref): Use the type of the name for printing the field name,
 instead of the name of the complete binding-ref expression.

* gnat2why/why/why-gen-terms.adb
(Has_Dereference_Or_Any_Or_Self): Consider ref-bindings like dereferences
 for determining the capability to express an expression in pure logic.

* QA04-034__reduced_volatile/test.out
* QA04-034__reduced_volatile/test.py
* QA04-034__reduced_volatile/vol.adb
* QA04-034__reduced_volatile/vol.ads
* QA04-034__reduced_volatile/vol2.adb
* QA04-034__reduced_volatile/vol2.ads
Add reduced test.

Change-Id: I763d181934642b54a851b8c1c7b1d47571a05aaa