Q510-009 enable an UOK test
When compiling this test gnatbind detects a circular elaboration; I hoped
that gnatprove can detect this when proving.

However, this it was too hard to reuse the old ABE, or
access-before-elaboration, mechanism from the frontend (because for
gnatprove it would have to be adapted to work with no-body-yet is present).
I don't think that the new ABE mechanism is any easier to reuse.

So let's enable this test with no action at the gnatprove side, but call
gnatmake to make sure that gnatbind will detect the problem.

* testsuite/gnatprove/tests/Q510-009__flow_circular_const/
Test with gnatmake; add output.

Change-Id: I1896a4b864a67def642a5c32b0f26d6fa9e3e425