NB14-019 bring test back to live
This test was raised because of wrong message:

  user.adb:7:26: high: "X" must be listed
                       in the Initializes aspect of "G" (SPARK RM 7.1.5(12))

After fixing it got rejected, because of generic IN actual instantiated with
variable input. In the meantime, it was disabled because of problems with
generic uncovered by the new generated globals.

Now, to my best knowledge, both generics and the generated globals are
working quite OK, so this test can be reenabled by removing the issue with
variable input.

* testsuite/gnatprove/tests/NB14-019__generic/
Reenable; add output.

Change-Id: Ieea5fc81040c4cb88d95aeec3329c78759b089db